requirements.txt
fastapi
uvicorn
paddleocr
pymupdf
opencv-python
sentence-transformers
faiss-cpu
spacy
